Moved utilities into separate folder
