--- functions/remote/prompt-defaults.js
@@ -1,4 +1,0 @@
-const validate = errMsg => input => !!input || errMsg
-const filter = input => input.trim()
-
-module.exports = { validate, filter }
